[REFACTOR] use fixtures.js in *.test.jsx
