+++ src/App.test.jsx
@@ -4,11 +4,7 @@ import { render } from '@testing-library/react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import regions from '../__fixtures__/regions';
-
-import categories from '../__fixtures__/categories';
-
-import restaurants from '../__fixtures__/restaurants';
+import { regions, categories, restaurants } from '../__fixtures__/fixtures';
 
 import App from './App';
